🚧 1K47B8 task: apply external agent result
