+++ .agentplane/tasks/202608292032-1K47B8/pr/diffstat.txt
@@ -2,12 +2,13 @@
  .../src/commands/shared/pr-meta/verify-log.test.ts |   60 ++
  .../src/commands/shared/pr-meta/verify-log.ts      |    4 +
  packages/core/src/tasks/index.ts                   |    2 +
- .../core/src/tasks/task-kernel/M1-QUALIFICATION.md |   91 ++
+ .../core/src/tasks/task-kernel/M1-QUALIFICATION.md |   93 ++
  packages/core/src/tasks/task-kernel/index.ts       |   23 +
  .../core/src/tasks/task-kernel/invariants.test.ts  |  427 ++++++++
  packages/core/src/tasks/task-kernel/invariants.ts  |  257 +++++
- packages/core/src/tasks/task-kernel/kernel.test.ts | 1021 +++++++++++++++++++
+ .../src/tasks/task-kernel/kernel.test-fixtures.ts  |  207 ++++
+ packages/core/src/tasks/task-kernel/kernel.test.ts |  833 +++++++++++++++
  packages/core/src/tasks/task-kernel/kernel.ts      | 1061 ++++++++++++++++++++
  packages/core/src/tasks/task-kernel/model.test.ts  |  108 ++
  packages/core/src/tasks/task-kernel/model.ts       |  375 +++++++
- 12 files changed, 3442 insertions(+)
+ 13 files changed, 3463 insertions(+)
