+++ .agentplane/tasks/202608292032-1K47B8/pr/meta.json
@@ -2,9 +2,9 @@
   "base": "main",
   "branch": "task/202608292032-1K47B8/implement-the-isolated-canonical-task-kernel",
   "created_at": "2026-08-29T20:33:30.307Z",
-  "diffstat_sha256": "sha256:894930b7a9abf0c5352a89824ba671bdbc80ee93c924e3e84e0a302a3287cc9b",
-  "last_verified_at": "2026-08-30T02:26:02.801Z",
-  "last_verified_diffstat_sha256": "sha256:ec11b8097dbf9eaa12bfbfe371a5d2555002a29cb2f2e08a13049e659dd7fa24",
+  "diffstat_sha256": "sha256:75f1a9ae077efe6d9b517513890e7358126dcc7f63095ec755e0153a1e80f2b7",
+  "last_verified_at": "2026-08-30T02:57:15.800Z",
+  "last_verified_diffstat_sha256": "sha256:894930b7a9abf0c5352a89824ba671bdbc80ee93c924e3e84e0a302a3287cc9b",
   "pr_number": 5877,
   "pr_url": "https://github.com/basilisk-labs/agentplane/pull/5877",
   "provider": {
@@ -20,6 +20,6 @@
   "task_id": "202608292032-1K47B8",
   "updated_at": "2026-08-29T22:04:21.476Z",
   "verify": {
-    "status": "pass"
+    "status": "fail"
   }
 }
