+++ .agentplane/tasks/202608292032-1K47B8/verification/20260830025715800-1f7012fe8ad951d0.json
@@ -1,0 +1,138 @@
+{
+  "details": "Command: bun run arch:check\nResult: pass\nEvidence: .agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-1\nScope: branch_pr task 202608292032-1K47B8 declared verification\n\nCommand: bunx vitest --config vitest.workspace.ts run packages/core/src/tasks/task-kernel/model.test.ts\nResult: pass\nEvidence: .agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-2\nScope: branch_pr task 202608292032-1K47B8 declared verification\n\nCommand: node .agentplane/policy/check-routing.mjs\nResult: pass\nEvidence: .agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-3\nScope: branch_pr task 202608292032-1K47B8 declared verification\n\nCommand: agentplane doctor\nResult: pass\nEvidence: .agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-4\nScope: branch_pr task 202608292032-1K47B8 declared verification\n\nCommand: bun run ci:local:full\nResult: fail\nEvidence: .agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-5\nScope: branch_pr task 202608292032-1K47B8 declared verification",
+  "digest": "sha256:1f7012fe8ad951d0919196b95efddce8c1f76b1524de2ef041c18b3c378a2241",
+  "implementation_sha": "f40d11ec3c70427e39ffe9ec784467e714e55076",
+  "input": {
+    "context": {
+      "digest": "sha256:6c5bf404e043b64f06d6fe8c0958e121fb33e8b689c974ad526d95c82071b672",
+      "paths": [
+        ".agentplane/WORKFLOW.md",
+        ".github/ISSUE_TEMPLATE/config.yml",
+        "bun.lock",
+        "depcruise.config.cjs",
+        "eslint.config.cjs",
+        "knip.json",
+        "package.json",
+        "packages/agentplane/package.json",
+        "packages/agentplane/src/cli/run-cli/commands/config.ts",
+        "packages/agentplane/src/runner/config.test.ts",
+        "packages/agentplane/src/runner/config.ts",
+        "packages/agentplane/tsconfig.json",
+        "packages/agentplane/tsup.config.ts",
+        "packages/core/package.json",
+        "packages/core/schemas/config.schema.json",
+        "packages/core/src/config/config.test.ts",
+        "packages/core/src/config/config.ts",
+        "packages/core/tsconfig.json",
+        "packages/core/tsup.config.ts",
+        "packages/recipes/package.json",
+        "packages/recipes/tsconfig.json",
+        "packages/recipes/tsup.config.ts",
+        "packages/spec/examples/config.json",
+        "packages/spec/package.json",
+        "packages/spec/schemas/config.schema.json",
+        "packages/testkit/package.json",
+        "packages/testkit/tsconfig.json",
+        "schemas/config.schema.json",
+        "tsconfig.json",
+        "turbo.json",
+        "vitest.config.ts",
+        "vitest.workspace.ts",
+        "website/bun.lock",
+        "website/docusaurus.config.ts",
+        "website/package.json",
+        "website/tsconfig.json"
+      ]
+    },
+    "digest": "sha256:0572849dc514a0e56b2d7f8bddf111bcb455d7f0b8b015e6761c716bbbe80b95",
+    "environment": {
+      "digest": "sha256:5eaa920b58022e097fc5b5afbe1e8caca7cdf6d6ad34230c2ac8eadaf77c86f9",
+      "runtime": {
+        "architecture": "arm64",
+        "bun_major": null,
+        "node_major": "26",
+        "platform": "darwin"
+      }
+    },
+    "evidence": {
+      "details_digest": "sha256:ef7b033b544fb4a486cdacd4f466bba0cec9647bd3e89084be9045893073c98a",
+      "digest": "sha256:765c47408514581fb3d94c3be87e7ac22c16eeb75d7da3d6d2c6475183ca8e66",
+      "references": [
+        {
+          "digest": "sha256:88d8ce7894e0187ef5f4a3432c2c6dfb37784fb37bcefef63958d4b890cf27a4",
+          "fragment": "check-1",
+          "path": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json",
+          "reference": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-1",
+          "source": "filesystem"
+        },
+        {
+          "digest": "sha256:88d8ce7894e0187ef5f4a3432c2c6dfb37784fb37bcefef63958d4b890cf27a4",
+          "fragment": "check-2",
+          "path": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json",
+          "reference": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-2",
+          "source": "filesystem"
+        },
+        {
+          "digest": "sha256:88d8ce7894e0187ef5f4a3432c2c6dfb37784fb37bcefef63958d4b890cf27a4",
+          "fragment": "check-3",
+          "path": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json",
+          "reference": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-3",
+          "source": "filesystem"
+        },
+        {
+          "digest": "sha256:88d8ce7894e0187ef5f4a3432c2c6dfb37784fb37bcefef63958d4b890cf27a4",
+          "fragment": "check-4",
+          "path": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json",
+          "reference": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-4",
+          "source": "filesystem"
+        },
+        {
+          "digest": "sha256:88d8ce7894e0187ef5f4a3432c2c6dfb37784fb37bcefef63958d4b890cf27a4",
+          "fragment": "check-5",
+          "path": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json",
+          "reference": ".agentplane/tasks/202608292032-1K47B8/supervision/declared-checks.json#check-5",
+          "source": "filesystem"
+        }
+      ]
+    },
+    "execution": {
+      "authoritative_task_source": "task_worktree",
+      "base_ref": "main",
+      "base_sha": "dbaf4bc2878eab7b50f8ea6d14179d8d91030159",
+      "digest": "sha256:fb139c692420b29df8188d2ce5aa7fb73a44703dacdae54f72da1952fd8358d4",
+      "primary_task_id": "202608292032-1K47B8",
+      "reason_codes": [
+        "agent_preferred_branch_pr",
+        "effect_public_api",
+        "repository_branch_pr_floor"
+      ],
+      "repository_mode": "branch_pr",
+      "requested_mode": "branch_pr",
+      "route_source": "execution_contract",
+      "selected_mode": "branch_pr",
+      "task_ids": [
+        "202608292032-1K47B8"
+      ]
+    },
+    "implementation": {
+      "base_sha": "dbaf4bc2878eab7b50f8ea6d14179d8d91030159",
+      "digest": "sha256:8bf3f7bdb7f9b28e87aede130d7d0d8e2ee8aee5e56e341c109691f82c466542",
+      "strategy": "branch_diff",
+      "target_sha": "f40d11ec3c70427e39ffe9ec784467e714e55076"
+    },
+    "kind": "task_verification_input",
+    "schema_version": 4,
+    "verification_contract_digest": "sha256:44e48b72dc9e096d3baa20d7ed533ad88cd9372c4a0e0b3ae51a867269e7b704",
+    "verify_steps_digest": "sha256:b2604d3034fe83c6549ba4aff559910e07842a0075180ad24a7b934589864e02"
+  },
+  "kind": "task_verification_record",
+  "note": "Rework: Declared check failed: bun run ci:local:full",
+  "recorded_at": "2026-08-30T02:57:15.800Z",
+  "result": "needs_rework",
+  "schema_version": 2,
+  "scope": "1. Run the kernel model, reducer and invariant tests and the real verification-child subprocess isolation regression. Confirm deterministic transitions, rejection behavior, receipt idempotency and unchanged parent environment.\n2. Run `bun run arch:check`. Confirm the kernel has no filesystem, process, Git, provider, CLI, backend or legacy imports.\n3. Run `bun run ci:local:full`. Require all checks to pass, including fast tests and typechecking.\n4. Correct the qualification receipt to reference exact implementation and immutable evidence identities. Distinguish historical completed WorkItems from current requalification and do not reuse obsolete plan authority.\n5. Require evaluator acceptance, exact-head hosted checks, merge and Task Hosted Close before M2 starts. Do not publish a release.",
+  "scope_digest": "sha256:b2604d3034fe83c6549ba4aff559910e07842a0075180ad24a7b934589864e02",
+  "task_id": "202608292032-1K47B8",
+  "verification_command": "agentplane verify 202608292032-1K47B8 --rework --by SUPERVISOR",
+  "verifier": "SUPERVISOR"
+}
